Logger fwd header
